Move conversions into their own file and fix centiliter conversion
